+++ core/src/test/java/com/ericsson/bss/cassandra/ecchronos/core/repair/TestTableRepairJob.java
@@ -28,11 +28,12 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
 import static org.mockito.Mockito.when;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Iterator;
-import java.util.List;
 import java.util.concurrent.TimeUnit;
 
 import org.junit.After;
@@ -175,11 +176,11 @@ public class TestTableRepairJob
     {
         // mock
         doReturn(true).when(myRepairStateSnapshot).canRepair();
-
+        mockRepairGroup(0L);
         assertThat(myRepairJob.runnable()).isTrue();
 
         verify(myRepairState, times(1)).update();
-        verify(myRepairStateSnapshot, times(1)).canRepair();
+        verify(myRepairStateSnapshot, times(2)).canRepair();
     }
 
     @Test
@@ -187,12 +188,12 @@ public class TestTableRepairJob
     {
         // mock
         doReturn(false).doReturn(true).when(myRepairStateSnapshot).canRepair();
-
+        mockRepairGroup(0L);
         assertThat(myRepairJob.runnable()).isFalse();
         assertThat(myRepairJob.runnable()).isTrue();
 
         verify(myRepairState, times(2)).update();
-        verify(myRepairStateSnapshot, times(2)).canRepair();
+        verify(myRepairStateSnapshot, times(3)).canRepair();
     }
 
     @Test
@@ -313,7 +314,7 @@ public class TestTableRepairJob
         VnodeRepairStates vnodeRepairStates = VnodeRepairStatesImpl
                 .newBuilder(ImmutableList.of(new VnodeRepairState(tokenRange, replicas, 1234L)))
                 .build();
-        ReplicaRepairGroup replicaRepairGroup = new ReplicaRepairGroup(replicas, vnodes);
+        ReplicaRepairGroup replicaRepairGroup = new ReplicaRepairGroup(replicas, vnodes, System.currentTimeMillis());
 
         RepairStateSnapshot repairStateSnapshot = RepairStateSnapshot.newBuilder()
                 .withReplicaRepairGroups(Collections.singletonList(replicaRepairGroup))
@@ -357,7 +358,7 @@ public class TestTableRepairJob
         ImmutableList<LongTokenRange> vnodes = ImmutableList.of(tokenRange);
 
         VnodeRepairStates vnodeRepairStates = VnodeRepairStatesImpl.newBuilder(ImmutableList.of(new VnodeRepairState(tokenRange, replicas, 1234L))).build();
-        ReplicaRepairGroup replicaRepairGroup = new ReplicaRepairGroup(replicas, vnodes);
+        ReplicaRepairGroup replicaRepairGroup = new ReplicaRepairGroup(replicas, vnodes, System.currentTimeMillis());
 
         RepairStateSnapshot repairStateSnapshot = RepairStateSnapshot.newBuilder()
                 .withReplicaRepairGroups(Collections.singletonList(replicaRepairGroup))
@@ -485,4 +486,68 @@ public class TestTableRepairJob
         assertThat(myRepairJob.getView().getProgress()).isEqualTo(0);
     }
 
+    public void testGetRealPriority()
+    {
+        long lastRepaired = System.currentTimeMillis();
+        doReturn(lastRepaired).when(myRepairStateSnapshot).lastCompletedAt();
+        doReturn(false).when(myRepairStateSnapshot).canRepair();
+        mockRepairGroup(lastRepaired);
+        assertThat(myRepairJob.getRealPriority()).isEqualTo(-1);
+
+        lastRepaired = System.currentTimeMillis() - (TimeUnit.DAYS.toMillis(RUN_INTERVAL_IN_DAYS) - TimeUnit.HOURS.toMillis(1));
+        doReturn(lastRepaired).when(myRepairStateSnapshot).lastCompletedAt();
+        doReturn(true).when(myRepairStateSnapshot).canRepair();
+        mockRepairGroup(lastRepaired);
+        assertThat(myRepairJob.getRealPriority()).isEqualTo(-1);
+
+        lastRepaired = System.currentTimeMillis() - (TimeUnit.DAYS.toMillis(RUN_INTERVAL_IN_DAYS));
+        doReturn(lastRepaired).when(myRepairStateSnapshot).lastCompletedAt();
+        doReturn(true).when(myRepairStateSnapshot).canRepair();
+        mockRepairGroup(lastRepaired);
+        assertThat(myRepairJob.getRealPriority()).isEqualTo(1);
+
+        lastRepaired = System.currentTimeMillis() - (TimeUnit.DAYS.toMillis(RUN_INTERVAL_IN_DAYS) + TimeUnit.HOURS.toMillis(1));
+        doReturn(lastRepaired).when(myRepairStateSnapshot).lastCompletedAt();
+        doReturn(true).when(myRepairStateSnapshot).canRepair();
+        mockRepairGroup(lastRepaired);
+        assertThat(myRepairJob.getRealPriority()).isEqualTo(2);
+    }
+
+    @Test
+    public void testGetRealPrioritySnapshotLastRepairedAtLowerThanRepairGroups()
+    {
+        long lastRepairedAtSnapshot = System.currentTimeMillis() - TimeUnit.DAYS.toMillis(14);
+        doReturn(lastRepairedAtSnapshot).when(myRepairStateSnapshot).lastCompletedAt();
+        doReturn(true).when(myRepairStateSnapshot).canRepair();
+        long firstRepairGroupLastRepairedAt = System.currentTimeMillis() - (TimeUnit.DAYS.toMillis(RUN_INTERVAL_IN_DAYS));
+        ReplicaRepairGroup firstReplicaRepairGroup = getRepairGroup(new LongTokenRange(1, 2), firstRepairGroupLastRepairedAt);
+        mockRepairGroup(firstReplicaRepairGroup);
+        assertThat(myRepairJob.getRealPriority()).isEqualTo(1);
+
+        long secondRepairGroupLastRepairedAt = System.currentTimeMillis() - (TimeUnit.DAYS.toMillis(RUN_INTERVAL_IN_DAYS) + TimeUnit.HOURS.toMillis(1));
+        ReplicaRepairGroup secondReplicaRepairGroup = getRepairGroup(new LongTokenRange(2, 3), secondRepairGroupLastRepairedAt);
+        mockRepairGroup(secondReplicaRepairGroup, firstReplicaRepairGroup);
+        assertThat(myRepairJob.getRealPriority()).isEqualTo(2);
+    }
+
+    private void mockRepairGroup(long lastRepairedAt)
+    {
+        mockRepairGroup(getRepairGroup(new LongTokenRange(1, 2), lastRepairedAt));
+    }
+
+    private void mockRepairGroup(ReplicaRepairGroup ...replicaRepairGroups)
+    {
+        List<ReplicaRepairGroup> repairGroups = new ArrayList<>();
+        for (ReplicaRepairGroup replicaRepairGroup : replicaRepairGroups)
+        {
+            repairGroups.add(replicaRepairGroup);
+        }
+        when(myRepairStateSnapshot.getRepairGroups()).thenReturn(repairGroups);
+    }
+
+    private ReplicaRepairGroup getRepairGroup(LongTokenRange range, long lastRepairedAt)
+    {
+        ImmutableSet<Node> replicas = ImmutableSet.of(mock(Node.class), mock(Node.class));
+        return new ReplicaRepairGroup(replicas, ImmutableList.of(range), lastRepairedAt);
+    }
 }